collector: separate check unwrap helpers
